+++ lib/everythingrb.rb
@@ -5,6 +5,7 @@ require "json"
 
 require_relative "everythingrb/version"
 require_relative "everythingrb/core/array"
+require_relative "everythingrb/core/enumerable"
 require_relative "everythingrb/core/hash"
 require_relative "everythingrb/core/module"
 require_relative "everythingrb/core/ostruct"