+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `Array#key_map` and `Array#dig_map` for mapping over `Hash`
 - Added `with_index:` keyword argument to `Array#join_map`. Defaults to `false`
+- Added `Enumerable#join_map`
 
 ### Changed